# apps/server/openapi.json
{"components": {"schemas": {"Attempt": {"properties": {"attempted_at": {"format": "date-time", "title": "Attempted At", "type": "string"}, "config_id": {"title": "Config Id", "type": "integer"}, "duration_ms": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Duration Ms"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "instance_ocid": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Instance Ocid"}, "message": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Message"}, "status": {"title": "Status", "type": "string"}}, "required": ["config_id", "status"], "title": "Attempt", "type": "object"}, "Body_create_credential_api_credentials_post": {"properties": {"fingerprint": {"title": "Fingerprint", "type": "string"}, "name": {"title": "Name", "type": "string"}, "passphrase": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Passphrase"}, "private_key": {"contentMediaType": "application/octet-stream", "title": "Private Key", "type": "string"}, "region": {"title": "Region", "type": "string"}, "tenancy_ocid": {"title": "Tenancy Ocid", "type": "string"}, "user_ocid": {"title": "User Ocid", "type": "string"}}, "required": ["name", "tenancy_ocid", "user_ocid", "fingerprint", "region", "private_key"], "title": "Body_create_credential_api_credentials_post", "type": "object"}, "ChannelCreate": {"properties": {"config": {"discriminator": {"mapping": {"discord": "#/components/schemas/DiscordConfig", "ntfy": "#/components/schemas/NtfyConfig", "slack": "#/components/schemas/SlackConfig", "telegram": "#/components/schemas/TelegramConfig"}, "propertyName": "type"}, "oneOf": [{"$ref": "#/components/schemas/DiscordConfig"}, {"$ref": "#/components/schemas/SlackConfig"}, {"$ref": "#/components/schemas/TelegramConfig"}, {"$ref": "#/components/schemas/NtfyConfig"}], "title": "Config"}, "enabled": {"default": true, "title": "Enabled", "type": "boolean"}, "name": {"title": "Name", "type": "string"}, "type": {"enum": ["discord", "slack", "telegram", "ntfy"], "title": "Type", "type": "string"}}, "required": ["name", "type", "config"], "title": "ChannelCreate", "type": "object"}, "ChannelRead": {"properties": {"config": {"additionalProperties": true, "title": "Config", "type": "object"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "enabled": {"title": "Enabled", "type": "boolean"}, "id": {"title": "Id", "type": "integer"}, "name": {"title": "Name", "type": "string"}, "type": {"title": "Type", "type": "string"}, "updated_at": {"format": "date-time", "title": "Updated At", "type": "string"}}, "required": ["id", "name", "type", "enabled", "config", "created_at", "updated_at"], "title": "ChannelRead", "type": "object"}, "ChannelUpdate": {"properties": {"config": {"discriminator": {"mapping": {"discord": "#/components/schemas/DiscordConfig", "ntfy": "#/components/schemas/NtfyConfig", "slack": "#/components/schemas/SlackConfig", "telegram": "#/components/schemas/TelegramConfig"}, "propertyName": "type"}, "oneOf": [{"$ref": "#/components/schemas/DiscordConfig"}, {"$ref": "#/components/schemas/SlackConfig"}, {"$ref": "#/components/schemas/TelegramConfig"}, {"$ref": "#/components/schemas/NtfyConfig"}], "title": "Config"}, "enabled": {"default": true, "title": "Enabled", "type": "boolean"}, "name": {"title": "Name", "type": "string"}, "type": {"enum": ["discord", "slack", "telegram", "ntfy"], "title": "Type", "type": "string"}}, "required": ["name", "type", "config"], "title": "ChannelUpdate", "type": "object"}, "ConfigCreate": {"properties": {"availability_domain": {"title": "Availability Domain", "type": "string"}, "boot_volume_gb": {"default": 50, "minimum": 1.0, "title": "Boot Volume Gb", "type": "integer"}, "channel_ids": {"items": {"type": "integer"}, "title": "Channel Ids", "type": "array"}, "credential_id": {"title": "Credential Id", "type": "integer"}, "image_ocid": {"title": "Image Ocid", "type": "string"}, "max_attempts": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Max Attempts"}, "memory_gb": {"default": 24, "minimum": 1.0, "title": "Memory Gb", "type": "integer"}, "name": {"title": "Name", "type": "string"}, "ocpus": {"default": 4, "minimum": 1.0, "title": "Ocpus", "type": "integer"}, "retry_interval_sec": {"default": 60, "minimum": 1.0, "title": "Retry Interval Sec", "type": "integer"}, "shape": {"default": "VM.Standard.A1.Flex", "title": "Shape", "type": "string"}, "ssh_public_key": {"title": "Ssh Public Key", "type": "string"}, "subnet_ocid": {"title": "Subnet Ocid", "type": "string"}}, "required": ["name", "credential_id", "image_ocid", "subnet_ocid", "availability_domain", "ssh_public_key"], "title": "ConfigCreate", "type": "object"}, "ConfigRead": {"properties": {"availability_domain": {"title": "Availability Domain", "type": "string"}, "boot_volume_gb": {"title": "Boot Volume Gb", "type": "integer"}, "channel_ids": {"items": {"type": "integer"}, "title": "Channel Ids", "type": "array"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "credential_id": {"title": "Credential Id", "type": "integer"}, "enabled": {"title": "Enabled", "type": "boolean"}, "id": {"title": "Id", "type": "integer"}, "image_ocid": {"title": "Image Ocid", "type": "string"}, "max_attempts": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Max Attempts"}, "memory_gb": {"title": "Memory Gb", "type": "integer"}, "name": {"title": "Name", "type": "string"}, "ocpus": {"title": "Ocpus", "type": "integer"}, "retry_interval_sec": {"title": "Retry Interval Sec", "type": "integer"}, "shape": {"title": "Shape", "type": "string"}, "ssh_public_key": {"title": "Ssh Public Key", "type": "string"}, "subnet_ocid": {"title": "Subnet Ocid", "type": "string"}, "updated_at": {"format": "date-time", "title": "Updated At", "type": "string"}}, "required": ["id", "name", "credential_id", "enabled", "shape", "ocpus", "memory_gb", "boot_volume_gb", "image_ocid", "subnet_ocid", "availability_domain", "ssh_public_key", "retry_interval_sec", "max_attempts", "channel_ids", "created_at", "updated_at"], "title": "ConfigRead", "type": "object"}, "ConfigUpdate": {"properties": {"availability_domain": {"title": "Availability Domain", "type": "string"}, "boot_volume_gb": {"minimum": 1.0, "title": "Boot Volume Gb", "type": "integer"}, "channel_ids": {"items": {"type": "integer"}, "title": "Channel Ids", "type": "array"}, "credential_id": {"title": "Credential Id", "type": "integer"}, "image_ocid": {"title": "Image Ocid", "type": "string"}, "max_attempts": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Max Attempts"}, "memory_gb": {"minimum": 1.0, "title": "Memory Gb", "type": "integer"}, "name": {"title": "Name", "type": "string"}, "ocpus": {"minimum": 1.0, "title": "Ocpus", "type": "integer"}, "retry_interval_sec": {"minimum": 1.0, "title": "Retry Interval Sec", "type": "integer"}, "shape": {"title": "Shape", "type": "string"}, "ssh_public_key": {"title": "Ssh Public Key", "type": "string"}, "subnet_ocid": {"title": "Subnet Ocid", "type": "string"}}, "required": ["name", "credential_id", "shape", "ocpus", "memory_gb", "boot_volume_gb", "image_ocid", "subnet_ocid", "availability_domain", "ssh_public_key", "retry_interval_sec"], "title": "ConfigUpdate", "type": "object"}, "CredentialRead": {"properties": {"created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "fingerprint": {"title": "Fingerprint", "type": "string"}, "has_passphrase": {"title": "Has Passphrase", "type": "boolean"}, "id": {"title": "Id", "type": "integer"}, "name": {"title": "Name", "type": "string"}, "region": {"title": "Region", "type": "string"}, "tenancy_ocid": {"title": "Tenancy Ocid", "type": "string"}, "user_ocid": {"title": "User Ocid", "type": "string"}}, "required": ["id", "name", "tenancy_ocid", "user_ocid", "fingerprint", "region", "has_passphrase", "created_at"], "title": "CredentialRead", "type": "object"}, "DiscordConfig": {"properties": {"type": {"const": "discord", "default": "discord", "title": "Type", "type": "string"}, "webhook_url": {"title": "Webhook Url", "type": "string"}}, "required": ["webhook_url"], "title": "DiscordConfig", "type": "object"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "title": "Detail", "type": "array"}}, "title": "HTTPValidationError", "type": "object"}, "ImageOption": {"properties": {"display_name": {"title": "Display Name", "type": "string"}, "ocid": {"title": "Ocid", "type": "string"}, "operating_system": {"title": "Operating System", "type": "string"}, "os_version": {"title": "Os Version", "type": "string"}}, "required": ["ocid", "display_name", "operating_system", "os_version"], "title": "ImageOption", "type": "object"}, "LogEntry": {"description": "Structured, persisted log record (PRD \u00a76, \u00a79.3).\n\nDoubles as the FastAPI read schema. ``DbLogHandler`` populates it by\nextracting the predefined context keys (``config_id``/``attempt_id``/\n``credential_id``) from ``record.__dict__``; anything else lands in\n``extra`` as a JSON string. ``exc_info`` carries the formatted traceback\nfor ERROR/CRITICAL records.", "properties": {"attempt_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Attempt Id"}, "config_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Config Id"}, "credential_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Credential Id"}, "exc_info": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Exc Info"}, "extra": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Extra"}, "id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Id"}, "level": {"title": "Level", "type": "string"}, "logger": {"title": "Logger", "type": "string"}, "message": {"title": "Message", "type": "string"}, "timestamp": {"format": "date-time", "title": "Timestamp", "type": "string"}}, "required": ["level", "logger", "message"], "title": "LogEntry", "type": "object"}, "LogPage": {"properties": {"has_more": {"title": "Has More", "type": "boolean"}, "items": {"items": {"$ref": "#/components/schemas/LogEntry"}, "title": "Items", "type": "array"}, "next_cursor": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Next Cursor"}}, "required": ["items", "next_cursor", "has_more"], "title": "LogPage", "type": "object"}, "LoginRequest": {"properties": {"password": {"title": "Password", "type": "string"}, "username": {"title": "Username", "type": "string"}}, "required": ["username", "password"], "title": "LoginRequest", "type": "object"}, "NtfyConfig": {"properties": {"priority": {"default": 3, "maximum": 5.0, "minimum": 1.0, "title": "Priority", "type": "integer"}, "server_url": {"title": "Server Url", "type": "string"}, "tags": {"items": {"type": "string"}, "title": "Tags", "type": "array"}, "token": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Token"}, "topic": {"title": "Topic", "type": "string"}, "type": {"const": "ntfy", "default": "ntfy", "title": "Type", "type": "string"}}, "required": ["server_url", "topic"], "title": "NtfyConfig", "type": "object"}, "SetupRequest": {"properties": {"password": {"minLength": 8, "title": "Password", "type": "string"}, "username": {"minLength": 3, "title": "Username", "type": "string"}}, "required": ["username", "password"], "title": "SetupRequest", "type": "object"}, "SetupStatusResponse": {"properties": {"needs_setup": {"title": "Needs Setup", "type": "boolean"}}, "required": ["needs_setup"], "title": "SetupStatusResponse", "type": "object"}, "SlackConfig": {"properties": {"type": {"const": "slack", "default": "slack", "title": "Type", "type": "string"}, "webhook_url": {"title": "Webhook Url", "type": "string"}}, "required": ["webhook_url"], "title": "SlackConfig", "type": "object"}, "SubnetOption": {"properties": {"cidr_block": {"title": "Cidr Block", "type": "string"}, "display_name": {"title": "Display Name", "type": "string"}, "ocid": {"title": "Ocid", "type": "string"}}, "required": ["ocid", "display_name", "cidr_block"], "title": "SubnetOption", "type": "object"}, "TelegramConfig": {"properties": {"bot_token": {"title": "Bot Token", "type": "string"}, "chat_id": {"title": "Chat Id", "type": "string"}, "type": {"const": "telegram", "default": "telegram", "title": "Type", "type": "string"}}, "required": ["bot_token", "chat_id"], "title": "TelegramConfig", "type": "object"}, "TestSendResponse": {"properties": {"error": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error"}, "ok": {"title": "Ok", "type": "boolean"}}, "required": ["ok"], "title": "TestSendResponse", "type": "object"}, "UserResponse": {"properties": {"username": {"title": "Username", "type": "string"}}, "required": ["username"], "title": "UserResponse", "type": "object"}, "ValidationError": {"properties": {"ctx": {"title": "Context", "type": "object"}, "input": {"title": "Input"}, "loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "title": "Location", "type": "array"}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}, "required": ["loc", "msg", "type"], "title": "ValidationError", "type": "object"}, "VerifyResponse": {"properties": {"error": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error"}, "ok": {"title": "Ok", "type": "boolean"}}, "required": ["ok"], "title": "VerifyResponse", "type": "object"}}}, "info": {"title": "OCI Ampere A1 Auto-Provisioner", "version": "0.1.0"}, "openapi": "3.1.0", "paths": {"/api/attempts": {"get": {"operationId": "list_attempts_api_attempts_get", "parameters": [{"in": "query", "name": "config_id", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Config Id"}}, {"in": "query", "name": "status", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Status"}}, {"in": "query", "name": "limit", "required": false, "schema": {"default": 50, "maximum": 500, "minimum": 1, "title": "Limit", "type": "integer"}}], "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Attempt"}, "title": "Response List Attempts Api Attempts Get", "type": "array"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "List Attempts", "tags": ["attempts"]}}, "/api/auth/login": {"post": {"operationId": "login_api_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Login", "tags": ["auth"]}}, "/api/auth/logout": {"post": {"operationId": "logout_api_auth_logout_post", "responses": {"204": {"description": "Successful Response"}}, "summary": "Logout", "tags": ["auth"]}}, "/api/auth/me": {"get": {"operationId": "me_api_auth_me_get", "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}, "description": "Successful Response"}}, "summary": "Me", "tags": ["auth"]}}, "/api/auth/setup": {"get": {"description": "Public: report whether the admin account still needs to be created.", "operationId": "setup_status_api_auth_setup_get", "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SetupStatusResponse"}}}, "description": "Successful Response"}}, "summary": "Setup Status", "tags": ["auth"]}, "post": {"description": "Public: create the single admin (first signup) and auto-login.\n\n409 ``setup_already_done`` if an admin already exists.", "operationId": "setup_api_auth_setup_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SetupRequest"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Setup", "tags": ["auth"]}}, "/api/channels": {"get": {"operationId": "list_channels_api_channels_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ChannelRead"}, "title": "Response List Channels Api Channels Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Channels", "tags": ["channels"]}, "post": {"operationId": "create_channel_api_channels_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChannelCreate"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChannelRead"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Create Channel", "tags": ["channels"]}}, "/api/channels/{channel_id}": {"delete": {"operationId": "delete_channel_api_channels__channel_id__delete", "parameters": [{"in": "path", "name": "channel_id", "required": true, "schema": {"title": "Channel Id", "type": "integer"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Delete Channel", "tags": ["channels"]}, "put": {"operationId": "update_channel_api_channels__channel_id__put", "parameters": [{"in": "path", "name": "channel_id", "required": true, "schema": {"title": "Channel Id", "type": "integer"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChannelUpdate"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChannelRead"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Update Channel", "tags": ["channels"]}}, "/api/channels/{channel_id}/test": {"post": {"operationId": "test_channel_api_channels__channel_id__test_post", "parameters": [{"in": "path", "name": "channel_id", "required": true, "schema": {"title": "Channel Id", "type": "integer"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TestSendResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Test Channel", "tags": ["channels"]}}, "/api/configs": {"get": {"operationId": "list_configs_api_configs_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ConfigRead"}, "title": "Response List Configs Api Configs Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Configs", "tags": ["configs"]}, "post": {"operationId": "create_config_api_configs_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigCreate"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigRead"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Create Config", "tags": ["configs"]}}, "/api/configs/{config_id}": {"delete": {"operationId": "delete_config_api_configs__config_id__delete", "parameters": [{"in": "path", "name": "config_id", "required": true, "schema": {"title": "Config Id", "type": "integer"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Delete Config", "tags": ["configs"]}, "put": {"operationId": "update_config_api_configs__config_id__put", "parameters": [{"in": "path", "name": "config_id", "required": true, "schema": {"title": "Config Id", "type": "integer"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigUpdate"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigRead"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Update Config", "tags": ["configs"]}}, "/api/configs/{config_id}/toggle": {"post": {"operationId": "toggle_config_api_configs__config_id__toggle_post", "parameters": [{"in": "path", "name": "config_id", "required": true, "schema": {"title": "Config Id", "type": "integer"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigRead"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Toggle Config", "tags": ["configs"]}}, "/api/credentials": {"get": {"operationId": "list_credentials_api_credentials_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/CredentialRead"}, "title": "Response List Credentials Api Credentials Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Credentials", "tags": ["credentials"]}, "post": {"operationId": "create_credential_api_credentials_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_create_credential_api_credentials_post"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CredentialRead"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Create Credential", "tags": ["credentials"]}}, "/api/credentials/{credential_id}": {"delete": {"operationId": "delete_credential_api_credentials__credential_id__delete", "parameters": [{"in": "path", "name": "credential_id", "required": true, "schema": {"title": "Credential Id", "type": "integer"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Delete Credential", "tags": ["credentials"]}}, "/api/credentials/{credential_id}/verify": {"post": {"operationId": "verify_credential_api_credentials__credential_id__verify_post", "parameters": [{"in": "path", "name": "credential_id", "required": true, "schema": {"title": "Credential Id", "type": "integer"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/VerifyResponse"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Verify Credential", "tags": ["credentials"]}}, "/api/logs": {"delete": {"operationId": "delete_logs_api_logs_delete", "parameters": [{"description": "delete records with timestamp < before", "in": "query", "name": "before", "required": true, "schema": {"description": "delete records with timestamp < before", "format": "date-time", "title": "Before", "type": "string"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Delete Logs", "tags": ["logs"]}, "get": {"operationId": "list_logs_api_logs_get", "parameters": [{"in": "query", "name": "levels", "required": false, "schema": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Levels"}}, {"description": "logger name prefix", "in": "query", "name": "logger", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "logger name prefix", "title": "Logger"}}, {"in": "query", "name": "config_id", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Config Id"}}, {"in": "query", "name": "since", "required": false, "schema": {"anyOf": [{"format": "date-time", "type": "string"}, {"type": "null"}], "title": "Since"}}, {"in": "query", "name": "until", "required": false, "schema": {"anyOf": [{"format": "date-time", "type": "string"}, {"type": "null"}], "title": "Until"}}, {"description": "message substring (LIKE)", "in": "query", "name": "q", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "message substring (LIKE)", "title": "Q"}}, {"in": "query", "name": "limit", "required": false, "schema": {"default": 50, "maximum": 200, "minimum": 1, "title": "Limit", "type": "integer"}}, {"description": "base64(last_id) \u2014 descending", "in": "query", "name": "cursor", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "base64(last_id) \u2014 descending", "title": "Cursor"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LogPage"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "List Logs", "tags": ["logs"]}}, "/api/logs/stream": {"get": {"description": "Live log stream over SSE (PRD \u00a78, \u00a79.3.7).\n\nSubscribes to the in-memory :data:`log_bus`, applies the same filter\nsemantics as the query endpoint, and emits a ``ping`` heartbeat every\n15 seconds so idle proxies don't drop the connection.", "operationId": "stream_logs_api_logs_stream_get", "parameters": [{"in": "query", "name": "levels", "required": false, "schema": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Levels"}}, {"in": "query", "name": "logger", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Logger"}}, {"in": "query", "name": "config_id", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Config Id"}}], "responses": {"200": {"content": {"application/json": {"schema": {}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Stream Logs", "tags": ["logs"]}}, "/api/meta/availability-domains": {"get": {"operationId": "list_availability_domains_api_meta_availability_domains_get", "parameters": [{"in": "query", "name": "credential_id", "required": true, "schema": {"title": "Credential Id", "type": "integer"}}], "responses": {"200": {"content": {"application/json": {"schema": {"items": {"type": "string"}, "title": "Response List Availability Domains Api Meta Availability Domains Get", "type": "array"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "List Availability Domains", "tags": ["meta"]}}, "/api/meta/images": {"get": {"operationId": "list_images_api_meta_images_get", "parameters": [{"in": "query", "name": "credential_id", "required": true, "schema": {"title": "Credential Id", "type": "integer"}}, {"in": "query", "name": "shape", "required": false, "schema": {"default": "VM.Standard.A1.Flex", "title": "Shape", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ImageOption"}, "title": "Response List Images Api Meta Images Get", "type": "array"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "List Images", "tags": ["meta"]}}, "/api/meta/subnets": {"get": {"operationId": "list_subnets_api_meta_subnets_get", "parameters": [{"in": "query", "name": "credential_id", "required": true, "schema": {"title": "Credential Id", "type": "integer"}}], "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/SubnetOption"}, "title": "Response List Subnets Api Meta Subnets Get", "type": "array"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "List Subnets", "tags": ["meta"]}}, "/healthz": {"get": {"description": "Public, unauthenticated health check.", "operationId": "healthz_healthz_get", "responses": {"200": {"content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "title": "Response Healthz Healthz Get", "type": "object"}}}, "description": "Successful Response"}}, "summary": "Healthz", "tags": ["meta"]}}}}